Reuse shared directory hashing for wheel archives
